fix: botones
