+++ .storybook/preview.ts
@@ -17,7 +17,7 @@ const headers = [
   'Componentes|Audio',
   'Componentes|Barra de progreso',
   'Componentes|Botones',
-  'Componentes|Botones/En caja',
+  'Componentes|Botones/Estados',
   'Componentes|Colapsable',
   'Componentes|Desplegables',
   'Componentes|Enlace',
@@ -57,6 +57,7 @@ const headers = [
   'Formulario|Texto',
   'Formulario|Validación',
   'Plantillas|Plantilla',
+  'Organismos|Botones',
   'Organismos|Footer (pie de página)',
   'Organismos|Header (encabezado)/Con buscador',
   'Organismos|Header (encabezado)/Sin buscador',
