+++ src/components/Button/Button.stories.tsx
@@ -16,53 +16,33 @@ export default {
   decorators: [withA11y]
 };
 
-export const ButtonSolid = (): JSX.Element => {
+export const ButtonTypes = (): JSX.Element => {
   return (
-    <div className="btn-wrapper">
-      {buttonTypes.map((item) => (
-        <Button key={item.type} type={item.type}>
-          {item.name}
-        </Button>
-      ))}
-      <a className="btn btn-primary" href="#" target="_blank">
-        Enlace
-      </a>
-    </div>
+    <>
+      <div className="btn-wrapper">
+        {buttonTypes.map((item) => (
+          <Button key={item.type} type={item.type}>
+            {item.name}
+          </Button>
+        ))}
+        <a className="btn btn-primary" href="#" target="_blank">
+          Enlace
+        </a>
+      </div>
+      <div className="btn-wrapper">
+        {buttonTypes.map((item) => (
+          <Button key={item.type} type={item.type} outline={true}>
+            {item.name}
+          </Button>
+        ))}
+        <a className="btn btn-outline-primary" href="#" target="_blank">
+          Enlace
+        </a>
+      </div>
+    </>
   );
 };
-ButtonSolid.story = { name: 'Colores' };
-
-export const ButtonFocus = (): JSX.Element => {
-  return (
-    <div className="btn-wrapper">
-      {buttonTypes.map((item) => (
-        <Button key={item.type} type={item.type} className="focus">
-          {item.name}
-        </Button>
-      ))}
-      <a className="btn btn-primary focus" href="#" target="_blank">
-        Enlace
-      </a>
-    </div>
-  );
-};
-ButtonFocus.story = { name: 'Con foco' };
-
-export const ButtonActive = (): JSX.Element => {
-  return (
-    <div className="btn-wrapper">
-      {buttonTypes.map((item) => (
-        <Button key={item.type} type={item.type} className="active">
-          {item.name}
-        </Button>
-      ))}
-      <a className="btn btn-primary active" href="#" target="_blank">
-        Enlace
-      </a>
-    </div>
-  );
-};
-ButtonActive.story = { name: 'Activos' };
+ButtonTypes.story = { name: 'Tipos' };
 
 export const ButtonSizes = (): JSX.Element => {
   return (
@@ -80,19 +60,6 @@ export const ButtonSizes = (): JSX.Element => {
 };
 ButtonSizes.story = { name: 'Tamaños' };
 
-export const ButtonDisabled = (): JSX.Element => {
-  return (
-    <div className="btn-wrapper">
-      {sizesButton.map((item) => (
-        <Button key={item.size} type="secondary" disabled={true} size={item.size}>
-          {item.name}
-        </Button>
-      ))}
-    </div>
-  );
-};
-ButtonDisabled.story = { name: 'Deshabilitados' };
-
 export const ButtonBlock = (): JSX.Element => {
   return (
     <div className="btn-block-wrapper">
@@ -104,35 +71,27 @@ export const ButtonBlock = (): JSX.Element => {
 };
 ButtonBlock.story = { name: 'Expandible' };
 
-export const ButtonOutline = (): JSX.Element => {
-  return (
-    <div className="btn-wrapper">
-      {buttonTypes.map((item) => (
-        <Button key={item.type} type={item.type} outline={true}>
-          {item.name}
-        </Button>
-      ))}
-      <a className="btn btn-outline-primary" href="#" target="_blank">
-        Enlace
-      </a>
-    </div>
-  );
-};
-ButtonOutline.story = { name: 'Con bordes' };
-
 export const ButtonWithIcon = (): JSX.Element => {
   return (
     <>
       <div className="btn-wrapper">
         {sizes.map((size) => (
-          <Button key={size} type="primary" size={size} iconBx="bxs-car">
+          <Button key={size} type="primary" size={size} iconBx="bxs-info-circle">
             Botón
           </Button>
         ))}
         {sizes.map((size) => (
-          <Button key={size} type="primary" outline={true} size={size} iconMaterial="drive_eta">
+          <Button key={size} type="primary" outline={true} size={size} iconMaterial="info">
             Botón
           </Button>
+        ))}
+      </div>
+      <div className="btn-wrapper">
+        {sizes.map((size) => (
+          <Button key={size} type="primary" size={size} iconBx="bxs-info-circle"></Button>
+        ))}
+        {sizes.map((size) => (
+          <Button key={size} type="primary" outline={true} size={size} iconMaterial="info"></Button>
         ))}
       </div>
     </>
