+++ src/organisms/Button/Button.stories.scss
@@ -1,0 +1,7 @@
+.btn-container {
+  max-width: 340px;
+
+  .card {
+    margin-bottom: 1rem;
+  }
+}
