+++ src/scss/components/_button.scss
@@ -196,14 +196,14 @@ a.btn {
   gap: 0.25rem;
 
   i,
-  span {
-    font-size: 0.9375rem;
+  .material-icons-round {
+    font-size: 1rem;
   }
 
   &.btn-lg {
     i,
-    span {
-      font-size: $btn-font-size;
+    .material-icons-round {
+      font-size: 1.25rem;
     }
   }
